requirements.txt
discord.py==0.16.6
requests==2.9.1